+++ Session/src/Session.php
@@ -31,12 +31,11 @@ class Session implements \IteratorAggregate
     {
         // Session is a singleton.
         if (isset(self::$instance)) {
-            throw new \RuntimeException("Session has already been initialized.", 500);
+            throw new \RuntimeException('Session has already been initialized.', 500);
         }
 
         // Destroy any existing sessions started with session.auto_start
-        if ($this->isSessionStarted())
-        {
+        if ($this->isSessionStarted()) {
             session_unset();
             session_destroy();
         }
@@ -52,12 +51,6 @@ class Session implements \IteratorAggregate
         register_shutdown_function([$this, 'close']);
         session_cache_limiter('nocache');
 
-        if (isset($this->count)) {
-            $this->count++;
-        } else {
-            $this->count = 1;
-        }
-
         self::$instance = $this;
     }
 
@@ -69,7 +62,7 @@ class Session implements \IteratorAggregate
      */
     public function instance()
     {
-        if (!isset(self::$instance)) {
+        if (null === self::$instance) {
             throw new \RuntimeException("Session hasn't been initialized.", 500);
         }
 
@@ -101,7 +94,7 @@ class Session implements \IteratorAggregate
     /**
      * Get session ID
      *
-     * @return string Session ID
+     * @return string|null Session ID
      */
     public function getId()
     {
@@ -126,7 +119,7 @@ class Session implements \IteratorAggregate
     /**
      * Get session name
      *
-     * @return string
+     * @return string|null
      */
     public function getName()
     {