+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# v1.4.0
+## 06/13/2018
+
+1. [](#new)
+    * `BlueprintForm`: Implemented support for multiple `import@`s and partial `import@`s (#17)
+1. [](#improved)
+    * `YamlFile`: Added support for `@data` without quoting it (fixes issues with Symfony 3.4 if `compat=true`)
+    * `YamlFile`: Added compatibility mode which falls back to Symfony YAML 2.8.38 if parsing with newer version fails
+    * `YamlFile`: Make `compat` and `native` settings global, enable `native` setting by default
+    * General code cleanup, some optimizations
+1. [](#bugfix)
+    * `Session`: Removed broken request counter
+
 # v1.3.9
 ## 10/08/2017
 
@@ -41,7 +54,7 @@
     * Fixed `IniFile::content()` should not fail if file doesn't exist
     * Session: Protection against invalid session cookie name throwing exception
     * Session: Do not destroy session on CLI
-    * BlueprintSchema: Fixed warning when field list is not what was expected 
+    * BlueprintSchema: Fixed warning when field list is not what was expected
 
 # v1.3.3
 ## 10/06/2016
@@ -68,9 +81,9 @@
     * Add new function UniformResourceLocator::fillCache()
 1. [](#bugfix)
     * Fix collections support in BluprintSchema::extra()
-    * Fix exception in stream wrapper when scheme is not defined in locator 
+    * Fix exception in stream wrapper when scheme is not defined in locator
     * Prevent UniformResourceLocator from resolving paths outside of defined scheme paths (#8)
-    * Fix breaking YAML files which start with three dashes (#5) 
+    * Fix breaking YAML files which start with three dashes (#5)
 
 # v1.3.0
 ## 03/07/2016
@@ -96,7 +109,7 @@
 ## 10/24/2015
 
 1. [](#new)
-    * **Backwards compatibility break**: Blueprints class needs to be initialized with `init()` if blueprints contain `@data-*` fields 
+    * **Backwards compatibility break**: Blueprints class needs to be initialized with `init()` if blueprints contain `@data-*` fields
     * Renamed NestedArrayAccess::remove() into NestedArrayAccess::undef() to avoid name clashes
 
 # v1.1.4